requirements.txt
python-telegram-bot==22.5
yfinance~=0.2.66
pandas~=2.3.3
numpy~=2.3.4
matplotlib~=3.10.7
scikit-learn~=1.7.2
statsmodels~=0.14.5
torch~=2.9.0
scipy~=1.16.3